requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
openai>=1.40
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9
dateparser>=1.2
httpx>=0.27
loguru>=0.7
pytest>=8.2
